Move node_identifiers_post to website/identifiers/views.py.
