+++ website/project/views/register.py
@@ -20,8 +20,8 @@ from website.project.decorators import (
     must_not_be_registration, must_be_registration,
     must_not_be_retracted_registration
 )
-from website.identifiers.utils import get_or_create_identifiers, build_ezid_metadata
-from osf.models import Identifier, MetaSchema, NodeLog
+from website.identifiers.utils import build_ezid_metadata
+from osf.models import Identifier, MetaSchema
 from website.project.utils import serialize_node
 from website.util.permissions import ADMIN
 from website import language
@@ -216,33 +216,6 @@ def osf_admin_change_status_identifier(node, status):
         doi, metadata = build_ezid_metadata(node)
         client = EzidClient(settings.EZID_USERNAME, settings.EZID_PASSWORD)
         client.change_status_identifier(status, doi, metadata)
-
-
-@must_be_valid_project
-@must_have_permission(ADMIN)
-def node_identifiers_post(auth, node, **kwargs):
-    """Create identifier pair for a node. Node must be a public registration.
-    """
-    if not node.is_public or node.is_retracted:
-        raise HTTPError(http.BAD_REQUEST)
-    if node.get_identifier('doi') or node.get_identifier('ark'):
-        raise HTTPError(http.BAD_REQUEST)
-    try:
-        identifiers = get_or_create_identifiers(node)
-    except HTTPError:
-        raise HTTPError(http.BAD_REQUEST)
-    for category, value in identifiers.iteritems():
-        node.set_identifier_value(category, value)
-    node.add_log(
-        NodeLog.EXTERNAL_IDS_ADDED,
-        params={
-            'parent_node': node.parent_id,
-            'node': node._id,
-            'identifiers': identifiers,
-        },
-        auth=auth,
-    )
-    return identifiers, http.CREATED
 
 
 def get_referent_by_identifier(category, value):
